get_db trasport to database and schema to schema file
