+++ Code/databse.py
@@ -4,3 +4,11 @@ from sqlalchemy.orm import sessionmaker
 DATABASE_URL = "sqlite:///./inventory.db"
 engine = create_engine(DATABASE_URL, connect_args={"check_same_thread": False})
 SessionLocal = sessionmaker(bind=engine, autocommit=False, autoflush=False)
+
+#make db func and sesion maker
+def get_db():
+    db = SessionLocal()
+    try:
+        yield db
+    finally:
+        db.close()
