+++ Code/main.py
@@ -6,6 +6,8 @@ from pydantic import Field
 from model import Supplier , Base
 from databse import engine , SessionLocal
 from typing import Optional
+from schema import SupplierCreate , SupplierResponse , SupplierUpdate
+from databse import get_db  
 
 
 #make engine
@@ -13,36 +15,6 @@ Base.metadata.create_all(bind=engine)
 
 #make app
 app = FastAPI()
-
-#make db func and sesion maker
-def get_db():
-    db = SessionLocal()
-    try:
-        yield db
-    finally:
-        db.close()
-
-#make acceptable value for supplier
-class SupplierCreate(BaseModel):
-    name: str
-    email: EmailStr
-    contact: str | None = None
-    delivery_days : int
-
-#make acceptable value for active or not
-class SupplierResponse(SupplierCreate):
-    id: int
-    is_active: bool
-
-    class Config:
-        orm_mode = True
-
-#for changing info
-class SupplierUpdate(BaseModel):
-    name: Optional[str] = Field(None)
-    email: Optional[EmailStr] = Field(None)
-    contact: Optional[str] = Field(None)
-    delivery_time_days: Optional[int] = Field(None)
 
 #add post
 @app.post("/suppliers/", response_model=SupplierResponse)
